agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
arize-phoenix==13.3.0
arize-phoenix-client==1.29.1
arize-phoenix-evals==2.10.0
arize-phoenix-otel==0.14.0
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.8
backoff==2.2.1
bcrypt==5.0.0
black==26.1.0
blockbuster==1.5.26
build==1.4.0
cachetools==7.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.1
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
croniter==6.0.0
cross-web==0.4.1
cryptography==46.0.5
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
executing==2.2.1
fastapi==0.133.1
filelock==3.24.3
filetype==1.2.0
flatbuffers==25.12.19
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
graphql-core==3.2.7
greenlet==3.3.2
groq==0.37.1
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-health-checking==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kubernetes==35.0.0
langchain==1.2.10
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-api==0.7.56
langgraph-checkpoint==4.0.0
langgraph-cli==0.4.12
langgraph-prebuilt==1.0.8
langgraph-runtime-inmem==0.26.0
langgraph-sdk==0.3.9
langsmith==0.7.7
ldap3==2.9.1
lia-web==0.3.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openinference-instrumentation==0.1.45
openinference-instrumentation-langchain==0.1.61
openinference-instrumentation-openai==0.1.41
openinference-semantic-conventions==0.1.26
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
platformdirs==4.9.2
posthog==5.4.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sqlean.py==3.50.4.5
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.52.1
strawberry-graphql==0.287.3
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
truststore==0.10.4
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
